add support for repeated range operator calls
Change the internal implementation for "expression&repetition"
operators to support this kind of mixfix operators.  The `'macro`
protocol was used only for postfix operators, but it actually couldn't
work if the expression handler did anything more with the
tail (because the expression handler couldn't know that it needed to
parse repetitions instead of expressions).  Instead, build in the
support for postfix (and "nofix") operators, and add mixfix modes to
support some mixfix range operators.