make syntax_class instances more object-like
When p is a pattern variable that is annotated with syntax class C,
then preserve the fields of p on the syntax-object value of p, instead
of allowing access to the fields only statically within the scope of
the binding. This change also fixes the problem where binding via
`pattern` did not preserve fields outside the `pattern` form.

A new `Syntax.matched_of` annotation constructor recognizes syntax
objects that are the result of matches to the syntax class C and
provides statically resolved access to the fields, which supports
passing a syntax object from a class-annotated match to a helper
function. Along similar lines, a field in a syntax class can be an
instance of another syntax class retains its nested fields.

A syntax class more like a veneer than a class in one way: it provides
access to fields that are repetitions, while those fields cannot be
accessed dynamically (since dynamic access of a repetition doesn't
work in general).

To implement a syntax object with fields, one strategy is to use
wrapper struct; then, the Rhombus `Syntax` annotation is either a
syntax object or a wrapped one. That strategy is best for checking
that fields are pruned at the right boundaries, but it creates some
cross-phase persistence issues, at least for error reporting. The
strategy enabled here is to attach fields through a syntax property,
instead.