make expression expansion less lazy
Formerly, when a Rhombus macro returned an expression that is not yet
parsed, then its expansion could be delayed until forced later. The
idea was that this laziness might allow more flexibility in expansion
order and variable references, but it doesn't apear to be useful ---
and it's counter-productive, because macros that use
`statinfo_meta.lookup` in principle all needed to use
`expr_meta.parse_more`, and practically none did. Making expansion
eager resolves the problem, apparently without losing any worthwhile
flexibliity.