add `group_option_sequence` and `term_option_sequence`
The `group_option_sequence` and `term_option_sequence` unquote-binding
forms acheieve an effect similar to using `~alts` with `~optional` in
`syntax-parse`. One difference is that when an option is used multiple
times, then an error is reported instead of failing to match. It's
possible that the fail-to-match behavior is also useful, in which case
we could add an option in the future.