define operator orders for rx and charset operators
Also, refine the core orders, including splitting `comparison` into
`order_comparison` and `equivalence`. The advamtage of the split is
that `logical_negation` can have a precedence relation to
`equivalence` but not to `order_comparison`, so `! 1 > 2` is a syntax
error; then again, it may be a bad idea to try to squeeze the job of a
type system into precedence relationships.