pict: improved model of identity
When a pict is rebuilt, give the rebuilt pict an identity that matches
the original pict. That way, when q has descendant p, p has descendant
r, and q is rebuilt by replacing r, then the rebuilt p can be found
within the rebuilt q.

This commit also changes the `~children` keyword argument of `animate`
and `rebuildable` to `~depencencies`.