remove `#%body`
The idea of an implicit form for `:` bodies seemed nice, but we
haven't found a use for it. It seems that `defn.sequence_macro` is a
better apporach for things where we thought that `#%body` might be
useful. Meanwhile, removing `#%body` is a useful simplification: macro
authors will not have to carefully carry context from an input `:` to
an expansion that also uses `:`.