rename statinfo operatiorns from "union"/"intersect" to "and"/"or"
The functions names `statinfo_meta.union` and
`statinfo_meta.intersect` made sense from a key-value map perspective,
but they were backwards from a union/intersection type perspective.
Using the names `statinfo_meta.and` and `statinfo_meta.or` avoids the
direct mismatch and encourages reasoning in terms of the `&&` and `||`
annotation operators.

Thanks to Robby for the naming suggestion.