Change `List.index` and `List.find_index`
The old `List.index` was like `List.find`, taking a predicate as an
argument. The new `List.index` that takes a value to locate (like
`List.contains`), which makes it more like `treelist-index-of` and
Python's `index` method on list.

The old `List.index` is renamed here to `List.find_index`.

Ditto for methods in `PairList` and `MutableList`.