add `Range.is_empty` and `Range.canonicalize`
`Range.is_empty` isn't exactly intuitive, because it recognizes empty
ranges "by definition", but this can be accomodated by the combination
with `Range.canonicalize`.

`Range.canonicalize` returns canonical forms with respect to the
discrete domain, which is integers in our case.  This can be
particularly useful in `substring`-like operations to more easily
consume range arguments.