Improve repetition syntax error message
The [Repetitions](https://docs.racket-lang.org/rhombus/repetition.html) documentation includes this example code:

```
> def five = 5
> [five, ...]
five: used with wrong repetition depth
  expected: 0
  actual: 1
```

In the case where the bound variable isn't a repetition at all, I found this error message confusing. So this pull request adds a special case for that which changes the message to this:

```
> [five, ...]
five: not a repetition binding
  used repetition depth: 1
```

This new message is only shown when the variable isn't a repetition and is used in any repetition position. In all other cases, the original error message is used.