optional binding of an escape identifier by `try`
An identifier written after `try` and before the body block is bound
as an escape continuation for reference within the body of the `try`
form and for use during the dynamic extent of the `try` form.

In other words, `try escape: <body>` is like `(let/ec escape <body>)`
in Racket.