definitions and expression at the end of `class` effectively follow it
The change addresses #669 (because `space.enforest`, etc., bodies are
treated like `class`) as well as the problem described in the "Class
Namespaces" section of the guide (which is currently section 4.8). It
doesn't add any new syntax, but instead adjusts the way that
definitions and expressions at the end of a `class`, `interface`,
`veneer`, `space.enforest`, or `meta_namespace` form are treated.
Definitions and expressions that do not have any clauses or
declaration afterward are handled after the definition of the
enclosing class, etc., instead of before.

For example, this definition (from the Guide) is still an error,
because `export` follows the definition of `origin`, and exports are
needed to fully determine the definition of `Posn`:

```
class Posn(x, y):
    def origin = Posn(0, 0)
    export:
      origin
```

But the example can be repaired by simplying moving the `export` form
to the start of the `class` body:

```
class Posn(x, y):
    export:
      origin
    def origin = Posn(0, 0)
```

This works because definitions and expressions that are not followed
by any clauses or declarations are effectively moved after the class
definition, instead of before.

This is a subtle rule. The idea is to still allow definitions within a
`class` to affect later class clauses, but to take advantage of the
fact that definitions after any clauses are not going to affect any
clauses, so they can be delayed. (What if the definition introduces a
new clause form that is later used within the same body? This problem
is avoided by binding `class_clause.macro` itself as a class clause.)
I expect that not much existing code will be affected, but the update
in this commit to "rhombus-draw-lib/rhombus/draw/private/dc.rhm"
illustrates how existng code might rely on tail definitions preceding
a class (or one of its clauses).

More generally, the idea is that moving a definition later is a
natural thing to try when things don't work, so when a Rhombus user
encouteres a problem, they are likely to discover the right pattern
without scrutinizing the documentation. To help a user more, the error
message is improved here when a class name is used as an annotation
too early, since that's seems to be the most common problem with a
previously confusing error message ("not defined" instead of "used
before it's ready").