shrubbery: add support for minor variations
In a pedagogic context, especially, it can be helpful to tweak
shrubbery rules in a way that makes them less general and regular, but
also less confusing. In Shplait, for example, students frequently
write `|~else` (with no space) or starts a new line in a pattern or
template with indentation before `$` to make it continue the previous
line. Adding a notion of (controlled) variants gives languages some
control over those fine details.

Also, regularize and document shrubbery tool support, such as the
syntax colorer and indenter.