Add `parser/lex` in a "rhombus-parser" package
The `parser/lex` library is mostly a wrapper for `parser-tools/lex`.
It uses a new `lex_pattern` space for lexer regexps, which are a
little differen than `rx`, but te same syntax is used as much as
possible. The `rx_charset` space for character sets is shared with
`rx`.

A few Rhombus bugs discovered along the way are also repaired here.