adjust `import` modifier rules
The conventions around import modifiers relied on an internal
distinction between ones that can take a block argument and ones that
don't. That distinction was not exposed, and it wasn't documented
correctly.

Remove the internal distinction and clarify the rules. This is a
backward-incompatible change, but the only pattern I've found broken
in existing code at has is `import: X expose: y`, which is now more
often and better written as `import: X.y` (but `import: X: expose y`
would also work, with or without `:` after `expose`).