revise rules for defined identifier/operator as a macro
Instead of a special rule for `$('$')`, rely on the fact that `($)`
works to define `$`. Meanwhile, do not require parentheses to define a
dotted name, so something like `defn.macro 'ns.f'` can define `f` as
an extension to `ns`. To avoid a `.` being treated aspart of a defined
name, either put the defined name in parentheses or use `$('.')` to
match a literal dot.