`values` reducer: support general pattern bindings
Generalize `values` for `for` so that it can use pattern-matching bindings:

```
for values([sum, product] = [0, 1]) (i in 1..=10):
  [sum + i, product * i]
```

Note that even though two variables are bound by the pattern, the
`for` result is still just a single list. That's hopefully as you'd
expect.

There's a related subtlety with converting annotations. Conversions
apply to the iteration body, but not the result:

```
for values(lst :: Listable.to_list = PairList[0, 0, 0]) (i in 1..=10):
  lst :: List // assert that conversion happened in match
  PairList[i, i, i]
// final result is a PairList, not a List
```

That behavior is because the final results are the values that would
be fed into one more iteration, and not the pattern bindings or
conversions created by binding to new values. Any predicates
associated with an annotation are applied to the final values,
however, so

```
for values(lst :: List = [0, 0, 0]) (i in 1..=1):
  PairList[i, i, i]
```

throws an exception that `PairList[1, 1, 1]` is not a `List`. Overall,
this turns out to be an improvement over the old behevior, which
incorrectly discarded conversions, even for the iteration body.

Aside from the correction for converting bindings, the new `values` is
backward-compatible with existing uses. Avoiding code duplication in
the expansion of `values` needs a small, backward-incompatible
generation of the reducer protocol. The generalization allows a
reducer to lift definitions outside the iteration. The `values` form
lifts binding matchers so they can be used for the iteration body and
for the overall results. The reducer generalization seems likely to be
useful for future reducers, too.