repairs for `~group` syntax classes and `$()`
Restrict matches to a `~group` syntax class to nonempty term sequences
(so the match is well-formed as a group). Make `$()` at the end of a
pattern indicate a `~group` pattern, and generally allow group syntax
classes to be used in a pattern group that ends with `$()` (which
fixes an inconsistency with `macro` patterns and other pattern
contexts).