make `<..` and `<..=` ranges satisfy `{Sequence,List}Range`
Previously, only `..` and `..=` ranges were sequences/listables, on
the grounds that such a range has a definite first element, namely the
included starting point.  (And, more importantly, because the
precedents, Rust and Racket, didn't include `<..`- or `<..=`-like
ranges.)  I worried that there might be confusion regarding the first
element of a `<..` or `<..=` range, but it seems that making the first
included integer the first element is fine.  Moreover, this means that
a `<..` or `<..=` range would produce the same sequence/list as the
canonicalized range.